% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxedSeq.R
\name{createTaxedSeq}
\alias{createTaxedSeq}
\title{createTaxedSeq}
\usage{
createTaxedSeq(sequences, taxonomy)
}
\arguments{
\item{sequences}{DNAStringSet object containing the DNA}

\item{taxonomy}{taxonomy datafraome (usually from the taxonomizr
package)}

\item{taxIDs}{NCBI taxonomy IDs (usually from the taxonomizr
package)}
}
\value{
a taxedSeq object
}
\description{
Create an object of class taxedSeq
}
\details{
An object of class taxedSeq can be compiled from sequences,
matching taxiIDs and a taxonomy data-frame, usually obtained from
the taxnomomizr package. This object class makes sure that the
sequence and contents can be subsetted in a correct synchronous
manner.
}
\author{
Emanuel Heitlinger
}
