% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.ENAdownload}
\alias{.ENAdownload}
\title{.ENAdownload}
\usage{
.ENAdownload(files, URLs, should_be)
}
\arguments{
\item{files}{path where files are stored}

\item{URLs}{URL from which each file is downloaded}

\item{should_be}{number of sequences that should be retrived from
these files}
}
\value{
an object of the class ENAdownload
}
\description{
Constructor function for the ENAdownload class
}
\details{
This constructs objects of the class ENAdownload
}
\author{
Emanuel Heitlinger
}
