% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.getENAsequenceCount}
\alias{.getENAsequenceCount}
\title{.getENAsequenceCount}
\usage{
.getENAsequenceCount(marker, data = "noncoding_release", freequery = "")
}
\arguments{
\item{marker}{The marker to be queried see https://www.ebi.ac.uk/ena/browse/marker-portal-rest}

\item{data}{the database to be queried usually "noncoding_release", "coding_release" see https://www.ebi.ac.uk/ena/browser/advanced-search}

\item{freequery}{terms in addtion to marker to be included included in the query (e.g. to limit the scope of the search to particular nodes in the taxonomy)}
}
\value{
the number of sequences found for this marker search
}
\description{
Get the number of ENA released sequences for a marker
}
\details{
Post an http request to ENA to count the number of available
sequences for a marker
}
\author{
Emanuel Heitlinger
}
