% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{getAccession4ENAname}
\alias{getAccession4ENAname}
\title{getAccession4ENAname}
\usage{
getAccession4ENAname(ENAnames, COI = FALSE)
}
\arguments{
\item{ENAnames}{the full name of the (fasta) sequence in the ENA
download}
}
\value{
A string with the (NCBI) accession for a sequence
}
\description{
Convert ENA names to (NCBI) accession names
}
\details{
Extract the NCBI accession name for sequences from the name
European Nucleotide Archive gives to downloaded (fasta) sequences
}
\author{
Emanuel Heitlinger
}
