% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_curate.R
\name{getBadSpecies}
\alias{getBadSpecies}
\title{getBadSpecies}
\usage{
getBadSpecies(
  taxonomy,
  badstring = c("uncultured", "sp\\\\.", "environmental", "sample", "unidentified")
)
}
\arguments{
\item{taxonomy}{from taxonomizr::getTaxonomy}

\item{badstring}{a vector of uniformative taxonomical
annotations. These will be treated as if no annotation was
available at all.}
}
\value{

}
\description{
Identify uniformative species annotations based on a taxonomy
table from taxonomizr.
}
\details{
Some sequences from databases have uninformative taxnomical
annotation. This function can be used to idetnify those sequences,
it returns a logical indicating whether the species taxonomy for
an entry was uniformative.
}
\author{
Emanuel Heitlinger
}
