% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_curate.R
\name{getBadTaxa}
\alias{getBadTaxa}
\title{getBadTaxa}
\usage{
getBadTaxa(
  taxonomy,
  fromN = 10,
  allowedNA = 3,
  badstring = c("uncultured", "sp\\\\.", "environmental", "sample", "unidentified")
)
}
\arguments{
\item{taxonomy}{from taxonomizr::getTaxonomy}

\item{fromN}{perform screening only for taxa with more abundant in
the dataset than fromN (set to 0 to disable).}

\item{allowedNA}{the number of meaningless annotations allowed
over all levels of the taxonomy}

\item{badstring}{a vector of uniformative taxonomical
annotations. These will be treated as if no annotation was
available at all.}
}
\value{
a logical vector indicating which of the input taxa
    belongs to a "bad" taxon.
}
\description{
Identify uninformative annotations ("bad" taxa) based on a taxonomy
table from taxonomizr
}
\details{
Some sequences from databases have uninformative taxnomical
annotation. This function can idetnify those sequences and returns
a logical indicating whether the taxonomy for an entry was
uniformative.
}
\author{
Emanuel Heitlinger
}
