% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getENAdownloads}
\alias{getENAdownloads}
\title{getENAdownloads}
\usage{
getENAdownloads(
  marker,
  downloadDir,
  data = "noncoding_release",
  paging = "100000",
  freequery = ""
)
}
\arguments{
\item{marker}{The marker to be queried see
https://www.ebi.ac.uk/ena/browse/marker-portal-rest}

\item{downloadDir}{Directory into which the downloaded sequences
are saved}

\item{data}{the database to be queried usually
"noncoding_release", "coding_release" see
https://www.ebi.ac.uk/ena/browser/advanced-search}

\item{paging}{Downloads are broken down in into chunks. This is
the chunk-size for this pageing.}

\item{freequery}{terms in addtion to marker to be included
included in the query (e.g. to limit the scope of the search
to particular nodes in the taxonomy)}
}
\value{
Returns an object of class ENAdownload. Thi is a list
    containing filenames (files), the URLs the files were
    downloaded from (URLs) and the number of expected sequences in
    the database that should be downloaded (should_be)
}
\description{
Download marker sequences from ENA.
}
\details{
ENA has indexed their databases using hidden markov models. This
allows the retrieval of marker sequences beyond what is given in
the sequence description (e.g. also subsequences from complete
genomes, etc.). This function downloads these marker sequences
}
\author{
Emanuel Heitlinger
}
