% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getFiles}
\alias{getFiles}
\title{getFiles}
\usage{
getFiles(ENAdownload)
}
\arguments{
\item{ENAdownload}{the ENAdownload object to be accessed}
}
\value{
a vector of filenames
}
\description{
Accessor function for ENAdownload class
}
\details{
Access the files in an ENAdownload
}
\author{
Emanuel Heitlinger
}
