% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align_taxwise.R
\name{getMatrices}
\alias{getMatrices}
\title{getMatrices}
\usage{
getMatrices(taxedSeq, mc.cores = 1)
}
\arguments{
\item{taxedSeq}{A taxedSeq object}

\item{mc.cores}{number of threads/processor used. This works only
on UNIX systems via the package parallel.}
}
\value{
a list of distance matrices
}
\description{
Get distance matrices from sequence aligments of each species in a dataset
}
\details{
Alignments are created for each species seperately, if the species
has multiple sequences in the dataset
}
\author{
Emanuel Heitlinger
}
