% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align_taxwise.R
\name{getOutliers}
\alias{getOutliers}
\title{getOutliers}
\usage{
getOutliers(matrices, DNAdata, cutoff = 0.1)
}
\arguments{
\item{matrices}{A list of distance matrices from getMatrices}

\item{DNAdata}{A XString set of DNA or RNA sequences or a taxedSeq object}

\item{cutoff}{percent difference in the distance matrix above
which a sequence is considered an outlier}

\item{taxonomy}{from taxonomizr::getTaxonomy}
}
\value{
a vector strings representing names of outlier sequences
    in the original dataset further than cutoff distance from the
    mayority of sequences of the same species
}
\description{
Get outlier sequences based on distance matrices for each species
}
\details{
Based on distance matrices for each species (see getMatrices)
}
\author{
Emanuel Heitlinger
}
