% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align_taxwise.R
\name{getSeqnames}
\alias{getSeqnames}
\title{getSeqnames}
\usage{
getSeqnames(taxedSeq)
}
\arguments{
\item{taxedSeq}{a taxedSeq object}
}
\value{
a vector of names
}
\description{
Get the sequences names from a taxedSeq object
}
\details{
Simply extracts the names from the DNAstringSet of a taxedSeq
object (or of the DNAStringSet directly)
}
\author{
Emanuel Heitlinger
}
