% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getShouldBe}
\alias{getShouldBe}
\title{getShouldBe}
\usage{
getShouldBe(ENAdownload)
}
\arguments{
\item{ENAdownload}{the ENAdownload object to be accessed}
}
\value{
a numeric indicating the number of expected sequences in
    an ENAdownload
}
\description{
Accessor function for ENAdownload class
}
\details{
Access the number of expected sequences in an ENAdownload
}
\author{
Emanuel Heitlinger
}
