% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align_taxwise.R
\name{getSubsequeces}
\alias{getSubsequeces}
\title{getSubsequeces}
\usage{
getSubsequeces(matrices, DNAdata, cutoff = 0, mc.cores = 1)
}
\arguments{
\item{matrices}{A list of distance matrices from getMatrices}

\item{DNAdata}{A XString set of DNA or RNA sequences or an object
of class taxedSeq}

\item{taxonomy}{from taxonomizr::getTaxonomy}
}
\value{
a vector strings representing names of subsequences in the
    original dataset shorter than other sequences for the same
    species
}
\description{
Get (shorter) subsequences relative to longer sequences for each
species in a taxnomically annotated dataset
}
\details{
Based on distance matrices for each species (see getMatrices) this
function identifies the longest unique sequence and returns
shorter subsequencs with (perfect, when cutoff=0) matches (and
hence terminal gaps) in distance matrices based on alignments of
each species in a dataset
}
\author{
Emanuel Heitlinger
}
