% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierTrain.R
\name{idTaxaTrainAndClean}
\alias{idTaxaTrainAndClean}
\title{idTaxaTrainAndClean}
\usage{
idTaxaTrainAndClean(taxedSeq, normalizeSize = 10, badRetain = 100)
}
\arguments{
\item{taxedSeq}{An object of class taxedSeq containing a DNAStringSet and 
matching taxonomy}

\item{normalizeSize}{integer indicating to use only this number of
sequences in case of mutliple sequnces for the same taxon}

\item{badRetain}{postiv integer indicating to repeat classifier
training until the number of "problem sequences" is lower than
this number}
}
\value{
A list with three elements: [1] the sequence the latest
    classifier is based on [2] the taxonomy for the sequences in
    the latest classifier [3] the latest classifier itself
}
\description{
Recursively train a IdTaxa classifier
}
\details{
This function is a wrapper around DECIPER's Learn Taxa. It
recursively trains the classifier removing problem sequences in
each iteration.
}
\author{
Emanuel Heitlinger
}
