% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isACGT}
\alias{isACGT}
\title{Checks a vector of character sequences for whether they are entirely ACGT.}
\usage{
isACGT(seqs)
}
\arguments{
\item{seqs}{A \code{character} of candidate DNA sequences.}
}
\value{
A \code{logical}. Whether or not each input character was ACGT only.
}
\description{
Checks a vector of character sequences for whether they are entirely ACGT.
}
