% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxedSeq.R
\name{subselectTaxedSeq}
\alias{subselectTaxedSeq}
\title{}
\usage{
subselectTaxedSeq(x, i)
}
\arguments{
\item{x}{object of class taxedSeq}

\item{i}{index usedf for subsetting}

\item{taxedSeq}{an object of class taxedSeq to be subsetted}
}
\value{
subsetted object of class taxedSeq
}
\description{
Subset taxedSeq objects
}
\details{
Subsetting an object of class taxedSeq
}
\author{
Emanuel Heitlinger
}
