% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_curate.R
\name{.checkTaxonomy}
\alias{.checkTaxonomy}
\title{.checkTaxonomy}
\usage{
.checkTaxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{from taxonomizr::getTaxonomy or a taxedSeq object
containing this taxonomy}
}
\value{
a taxonomy matrix
}
\description{
Internal function checking the content of a taxonomy matrix
}
\details{
This internal function extracts the taxonomy table from a taxedSeq
object and checks whether the rquired species column is given in a
taxonomy table.
}
\author{
Emanuel Heitlinger
}
