% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatTaxedSeq.R, R/createTaxedSeq.R
\name{print.taxedSeq}
\alias{print.taxedSeq}
\title{print.taxedSeq}
\usage{
\method{print}{taxedSeq}(taxedSeq)

\method{print}{taxedSeq}(taxedSeq)
}
\arguments{
\item{taxedSeq}{an object of class taxedSeq to be printed}
}
\value{
printing to the console

printing to the console
}
\description{
Print the contents of a taxedSeq object

Print the contents of a taxedSeq object
}
\details{
Prints the DNAStingSet using the default and head an tail of the taxonomy

Prints the DNAStingSet using the default and head an tail of the taxonomy
}
\author{
Emanuel Heitlinger

Emanuel Heitlinger
}
